+++ compiler-utils/src/main/java/com/squareup/anvil/compiler/internal/reference/ClassReference.kt
@@ -66,8 +66,49 @@ public sealed class ClassReference : Comparable<ClassReference>, AnnotatedRefere
   public val packageFqName: FqName get() = classId.packageFqName
 
   public abstract val constructors: List<MemberFunctionReference>
+
+  @Deprecated(
+    "renamed to `declaredMemberFunctions`.  " +
+      "Use `memberFunctions` to include inherited functions.",
+    replaceWith = ReplaceWith("declaredMemberFunctions"),
+  )
   public abstract val functions: List<MemberFunctionReference>
+
+  /**
+   * All functions that are declared in this class, including overrides.
+   * This list does not include inherited functions that are not overridden by this class.
+   */
+  public abstract val declaredMemberFunctions: List<MemberFunctionReference>
+
+  /**
+   * All functions declared in this class or any of its super-types.
+   */
+  public val memberFunctions: List<MemberFunctionReference> by lazy(NONE) {
+    declaredMemberFunctions + directSuperTypeReferences()
+      .flatMap { it.asClassReference().memberFunctions }
+  }
+
+  @Deprecated(
+    "renamed to `declaredMemberProperties`.  \n" +
+      "Use `memberProperties` to include inherited properties.",
+    replaceWith = ReplaceWith("declaredMemberProperties"),
+  )
   public abstract val properties: List<MemberPropertyReference>
+
+  /**
+   * All properties that are declared in this class, including overrides.
+   * This list does not include inherited properties that are not overridden by this class.
+   */
+  public abstract val declaredMemberProperties: List<MemberPropertyReference>
+
+  /**
+   * All properties declared in this class or any of its super-types.
+   */
+  public val memberProperties: List<MemberPropertyReference> by lazy(NONE) {
+    declaredMemberProperties + directSuperTypeReferences()
+      .flatMap { it.asClassReference().memberProperties }
+  }
+
   public abstract val typeParameters: List<TypeParameterReference>
 
   protected abstract val innerClassesAndObjects: List<ClassReference>
@@ -146,7 +187,13 @@ public sealed class ClassReference : Comparable<ClassReference>, AnnotatedRefere
       clazz.containingFileAsJavaFile()
     }
 
-    override val functions: List<MemberFunctionReference.Psi> by lazy(NONE) {
+    @Deprecated(
+      "renamed to `declaredMemberFunctions`.  Use `memberFunctions` to include inherited functions.",
+      replaceWith = ReplaceWith("declaredMemberFunctions"),
+    )
+    override val functions: List<MemberFunctionReference.Psi> get() = declaredMemberFunctions
+
+    override val declaredMemberFunctions: List<MemberFunctionReference.Psi> by lazy(NONE) {
       clazz
         .children
         .filterIsInstance<KtClassBody>()
@@ -158,7 +205,13 @@ public sealed class ClassReference : Comparable<ClassReference>, AnnotatedRefere
       clazz.annotationEntries.map { it.toAnnotationReference(this, module) }
     }
 
-    override val properties: List<MemberPropertyReference.Psi> by lazy(NONE) {
+    @Deprecated(
+      "renamed to `declaredMemberProperties`.  \nUse `memberProperties` to include inherited properties.",
+      replaceWith = ReplaceWith("declaredMemberProperties"),
+    )
+    override val properties: List<MemberPropertyReference.Psi> get() = declaredMemberProperties
+
+    override val declaredMemberProperties: List<MemberPropertyReference.Psi> by lazy(NONE) {
       buildList {
         // Order kind of matters here, since the Descriptor APIs will list body/member properties
         // before the constructor properties.
@@ -263,9 +316,14 @@ public sealed class ClassReference : Comparable<ClassReference>, AnnotatedRefere
         )
     }
 
-    override val functions: List<MemberFunctionReference.Descriptor> by lazy(NONE) {
+    @Deprecated(
+      "renamed to `declaredMemberFunctions`.  Use `memberFunctions` to include inherited functions.",
+      replaceWith = ReplaceWith("declaredMemberFunctions"),
+    )
+    override val functions: List<MemberFunctionReference.Descriptor> get() = declaredMemberFunctions
+    override val declaredMemberFunctions: List<MemberFunctionReference.Descriptor> by lazy(NONE) {
       clazz.unsubstitutedMemberScope
-        .getContributedDescriptors(kindFilter = DescriptorKindFilter.FUNCTIONS)
+        .getDescriptorsFiltered(kindFilter = DescriptorKindFilter.FUNCTIONS)
         .filterIsInstance<FunctionDescriptor>()
         .filterNot { it is ConstructorDescriptor }
         .map { it.toFunctionReference(this) }
@@ -275,14 +333,19 @@ public sealed class ClassReference : Comparable<ClassReference>, AnnotatedRefere
       clazz.annotations.map { it.toAnnotationReference(this, module) }
     }
 
-    override val properties: List<MemberPropertyReference.Descriptor> by lazy(NONE) {
+    @Deprecated(
+      "renamed to `declaredMemberProperties`.  \nUse `memberProperties` to include inherited properties.",
+      replaceWith = ReplaceWith("declaredMemberProperties"),
+    )
+    override val properties: List<MemberPropertyReference.Descriptor>
+      get() = declaredMemberProperties
+
+    override val declaredMemberProperties: List<MemberPropertyReference.Descriptor> by lazy(NONE) {
       clazz.unsubstitutedMemberScope
         .getDescriptorsFiltered(kindFilter = DescriptorKindFilter.VARIABLES)
         .filterIsInstance<PropertyDescriptor>()
-        .filter {
-          // Remove inherited properties that aren't overridden in this class.
-          it.kind == DECLARATION
-        }
+        // Remove inherited properties that aren't overridden in this class.
+        .filter { it.kind == DECLARATION }
         .map { it.toPropertyReference(this) }
     }